# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Wiki API",
    "version": "0.2.0",
    "description": "Wiki MVP public API: auth, users, spaces, documents, task and phase links, evidence, attachments, templates, audit and search."
  },
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WikiLoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WikiAuthResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "logout",
        "security": [
          {
            "bearer": []
          }
        ],
        "responses": {
          "204": {
            "description": "Logged out"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/users/me": {
      "get": {
        "tags": [
          "users"
        ],
        "operationId": "getCurrentUser",
        "security": [
          {
            "bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WikiUserResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
    "/api/v1/spaces": {
      "get": {
        "tags": [
          "spaces"
        ],
        "operationId": "listSpaces",
        "security": [
          {
            "bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Spaces",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpaceListResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "post": {
        "tags": [
          "spaces"
        ],
        "operationId": "createSpace",
        "security": [
          {
            "bearer": []
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/CreateSpaceRequest"
        },
        "responses": {
          "201": {
            "description": "Space created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpaceResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/spaces/{space_key}": {
      "get": {
        "tags": [
          "spaces"
        ],
        "operationId": "getSpace",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/SpaceKey"
          }
        ],
        "responses": {
          "200": {
            "description": "Space",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpaceResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "put": {
        "tags": [
          "spaces"
        ],
        "operationId": "updateSpace",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/SpaceKey"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/UpdateSpaceRequest"
        },
        "responses": {
          "200": {
            "description": "Space updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpaceResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
    "/api/v1/spaces/{space_key}/members": {
      "get": {
        "tags": [
          "spaces"
        ],
        "operationId": "listSpaceMembers",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/SpaceKey"
          }
        ],
        "responses": {
          "200": {
            "description": "Space members",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpaceMemberListResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/spaces/{space_key}/members/{user_id}": {
      "put": {
        "tags": [
          "spaces"
        ],
        "operationId": "upsertSpaceMember",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/SpaceKey"
          },
          {
            "$ref": "#/components/parameters/UserId"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/UpsertSpaceMemberRequest"
        },
        "responses": {
          "200": {
            "description": "Space member upserted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpaceMemberResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "delete": {
        "tags": [
          "spaces"
        ],
        "operationId": "deleteSpaceMember",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/SpaceKey"
          },
          {
            "$ref": "#/components/parameters/UserId"
          }
        ],
        "responses": {
          "204": {
            "description": "Space member removed"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/spaces/{space_key}/tree": {
      "get": {
        "tags": [
          "documents"
        ],
        "operationId": "getSpaceTree",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/SpaceKey"
          }
        ],
        "responses": {
          "200": {
            "description": "Page tree",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpaceTreeResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/spaces/{space_key}/documents": {
      "post": {
        "tags": [
          "documents"
        ],
        "operationId": "createDocument",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/SpaceKey"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/CreateDocumentRequest"
        },
        "responses": {
          "201": {
            "description": "Document created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/documents/{document_id}": {
      "get": {
        "tags": [
          "documents"
        ],
        "operationId": "getDocument",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/DocumentId"
          }
        ],
        "responses": {
          "200": {
            "description": "Document",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/documents/{document_id}/draft": {
      "put": {
        "tags": [
          "documents"
        ],
        "operationId": "updateDocumentDraft",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/DocumentId"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/UpdateDocumentDraftRequest"
        },
        "responses": {
          "200": {
            "description": "Draft updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/documents/{document_id}/publish": {
      "post": {
        "tags": [
          "documents"
        ],
        "operationId": "publishDocument",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/DocumentId"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/PublishDocumentRequest"
        },
        "responses": {
          "200": {
            "description": "Document revision published",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentRevisionResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/documents/{document_id}/archive": {
      "post": {
        "tags": [
          "documents"
        ],
        "operationId": "archiveDocument",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/DocumentId"
          }
        ],
        "responses": {
          "200": {
            "description": "Document archived",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "responses": {
      "Error": {
        "description": "Error response",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    },
...
    "schemas": {
...
      "WikiAuthResponse": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string"
          },
          "refresh_token": {
            "type": "string"
          },
          "token_type": {
            "type": "string"
          },
          "user_id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "format": "password"
          },
          "display_name": {
            "type": "string"
          },
          "expires_in": {
            "type": "integer"
          }
        },
        "required": [
          "access_token",
          "refresh_token",
          "token_type",
          "user_id",
          "email",
          "username",
          "display_name",
          "expires_in"
        ]
      },
...
      "WikiLoginRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "format": "password"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
...
      "SpaceListResponse": {
        "type": "object",
        "properties": {
          "spaces": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpaceResponse"
            }
          }
        },
        "required": [
          "spaces"
        ]
      },
...
    }
  },
...
}